자유게시판 수정, 로그인 수정

+++ jobcruit/src/main/webapp/WEB-INF/views/job/free/detail.jsp
@@ -2,6 +2,7 @@
     pageEncoding="UTF-8"%>
 <%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core" %>
 <%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt" %>    
+<%@ taglib prefix="fn" uri="http://java.sun.com/jsp/jstl/functions" %>
 <!DOCTYPE html>
 <html>
 	<head>
@@ -41,11 +42,21 @@
 								    		<div>
 								    			<pre><c:out value="${free.content}" /></pre>
 								    		</div>
+							    			<ul class="uploadList">
+							    				<c:forEach items="${attach}" var="att">
+							    				
+							    				<li><div><span data-src="${att.fileName}">${fn:substringAfter(att.fileName, "_")}</span></div></li>
+							    				</c:forEach>
+							    			</ul>
 								    	
+								    		<form id="oper"></form>
+								    		
 												<div class="col-lg-12 form-actions no-margin-bottom text-right">
-													<button type="button" class="btn btn-default commentBtn"><span class="glyphicon glyphicon-chevron-down"></span>댓글</button>
-													<input type="button" value="수정/삭제" class="btn btn-primary editBtn">
-													<input type="button" value="취소" class="btn btn-primary closeBtn">
+													<button type="button" class="btn btn-default commentBtn"><span></span>댓글</button>
+													<c:if test="${free.mno eq login}">
+														<input type="button" value="수정/삭제" class="btn btn-primary editBtn">
+													</c:if>
+													<input type="button" value="닫기" class="btn btn-primary closeBtn">
 												</div>
 								    	</div>
 								    	<!-- /.boardDiv -->
@@ -54,16 +65,10 @@
 								    		<div class="commentWrite">
 								    			<table class="table">
 								    			<tr>
-								    				<td><textarea name="content" class="form-control" rows="3"></textarea></td>
-								    				<td><input type="submit" value="등록" class="btn btn-primary commRegBtn"></td>
+								    				<td class="col-lg-11"><textarea name="content" class="form-control" rows="3"></textarea></td>
+								    				<td class="col-lg-1"><input type="submit" value="등록" class="btn btn-primary commRegBtn"></td>
 								    			</tr>
 								    			</table>
-									   			<!-- <div class="col-lg-10 form-group">
-												    <textarea name="content" class="form-control" rows="3"></textarea>
-													</div>
-					                <div class="col-lg-2 form-actions no-margin-bottom text-right">
-		                      	<input type="submit" value="등록" class="btn btn-primary commRegBtn">
-		                    	</div> -->
 										    </div>      
 										    
 								    		<div class="commentList">
@@ -91,17 +96,17 @@
 											<form class="form-horizontal" action="/job/free/modify" method="post">
 						           	<input type="hidden" name="fno" value="${free.fno}">
 						              <div class="form-group">
-						                 <label class="control-label col-lg-4">제목</label>
-						                 <div class="col-lg-8">
+						                 <label class="control-label col-lg-2">제목</label>
+						                 <div class="col-lg-10">
 						                   <input type="text" name="title" placeholder="제목을 입력하세요" class="form-control" value='<c:out value="${free.title}" />'>
 						                 </div>
 						              </div>
 						              <!-- /.form-group -->
 						
 						              <div class="form-group">
-						                 <label class="control-label col-lg-4">내용</label>
-						                 <div class="col-lg-8">
-						                   <textarea name="content" class="form-control" rows="15"><c:out value="${free.content}" /></textarea>
+						                 <label class="control-label col-lg-2">내용</label>
+						                 <div class="col-lg-10">
+						                   <textarea name="content" placeholder="내용을 입력하세요" class="form-control" rows="15"><c:out value="${free.content}" /></textarea>
 						                 </div>
 						              </div>
 						              <!-- /.form-group -->
@@ -123,17 +128,13 @@
 				</div>
 			</div>
 		</div>
-		<%-- 
-		<form id="actionForm" action="/job/free/update" method="post">
-			<input type="hidden" name="fno" value="${free.fno}">
-		</form>
-		 --%>
+		
 		<%@include file="../common/footer.jsp" %> 
 		<%@include file="../common/bodyRes.jsp" %> 
 		
 		<script>
 		$(document).ready(function(){
-			//getCommentList();
+			//makeAttachList();
 		});
 		//===========================조회 화면==============================
 		$actionForm = $("#actionForm");
@@ -159,18 +160,43 @@
 				getCommentList();
 			}
 		});
+		
+		function makeAttachList(){
+			var attList = '${attach}';
+			console.log(attList);
+			if(attList != ""){
+				var attStr = "";
+				for(var i = 0;i<attList.length;i++){
+					console.log(attList[i]);
+					attStr += '<li><div>'+attList[i].fileName.split("_")[1]+'<span class="glyphicon glyphicon-remove" data-src="'+attList[i].fileName+'"></span></div></li>';
+				}
+				$(".uploadList").html(attStr);
+			}
+		}
+		
+		$(".uploadList").on("click", "span", function(e){
+			e.stopPropagation();	//안전한 코딩을 위한...
+			
+			var fileName = $(this).data("src");
+			var $oper = $("#oper");
+			$oper.attr("action","/job/upload/download/free");
+			$oper.html("<input type='hidden' name='name' value='"+fileName+"'>");
+			$oper.submit();
+		});
+	
 		//==========================조회 - 댓글 영역==========================
 		
 		var $commentList = $(".commentList");
 		$(".commRegBtn").on("click", function(e){
 			var $content = $(".commentWrite").find("textarea[name='content']");
+			
 			$.ajax({
 				url:"/job/free/comm",
 				type:"post",
 				data:{
 					fno:'${free.fno}',
 					mno:'${login}',
-					content:$content.val()
+					content:$content.val().replace(/\r\n|\n|\r/g, '<br />')
 				},
 				dataType:"text"
 			}).done(function(result){
@@ -185,6 +211,7 @@
 			
 			var $cno = $(this).attr("href");
 			var $content = $("#cno_"+$cno).find("textarea[name='content']");
+			
 			$.ajax({
 				url:"/job/free/comm/"+$cno,
 				type:"put",
@@ -193,7 +220,7 @@
 					"X-HTTP-Method-Override" : "PUT"
 				},
 				data : JSON.stringify({
-					content:$content.val()
+					content:$content.val().replace(/\r\n|\n|\r/g, '<br />')
 				}),
 				dataType:"text"
 			}).done(function(result){
@@ -215,22 +242,22 @@
 		  e.preventDefault();
 		  
 		  var $cno = $(this).attr("href");
-		  
-		  $.ajax({
-			  url:"/job/free/comm/"+$cno,
-			  type:"delete",
-			  dataType:"text"
-			}).done(function(result){
-				console.log(result);			
-				if(result == "success"){
-					showAlert('댓글이 삭제되었습니다');
-					getCommentList();
-				}else if(result == "fail"){
-					showAlert('댓글 삭제중 오류가 발생하였습니다','fail');
-				}						
-			});
+		  showConfirm("해당 댓글을 삭제하시겠습니까?", function($cno){
+			  $.ajax({
+				  url:"/job/free/comm/"+$cno,
+				  type:"delete",
+				  dataType:"text"
+				}).done(function(result){
+					if(result == "success"){
+						showAlert('댓글이 삭제되었습니다');
+						getCommentList();
+					}else if(result == "fail"){
+						showAlert('댓글 삭제중 오류가 발생하였습니다','error');
+					}						
+				});
+		  }, $cno);		  
 		});
-	
+		
 		//댓글 리스트 가져오기
 		function getCommentList(){
 			$.ajax({
@@ -238,7 +265,6 @@
 				data:"fno=${free.fno}",
 				dataType:"json"
 			}).done(function(commList){
-				console.log(commList);
 				if(commList){
 					makeComment(commList);
 				}
@@ -255,7 +281,7 @@
 					        '    	<b><span class="glyphicon glyphicon-user"></span>'+commList[i].mname+'</b>'+
 					        '			<span class="text-muted time">'+getDateTime(commList[i].regDate)+'</span>'+
 					       	' 	 </div>'+
-					       	'		<div class="pull-right button">'+
+					       	'		<div class="pull-right button'+(commList[i].mno == '${login}'?'':' hidden')+'">'+
 					       	'			<a href="'+commList[i].freeCno+'" class="commSaveBtn hidden"><span>저장</span></a>'+
 					       	'			<a href="'+commList[i].freeCno+'" class="commModiBtn"><span>수정</span></a>'+
 					       	'			<a href="'+commList[i].freeCno+'" class="commDelBtn"><span>삭제</span></a>'+
@@ -271,7 +297,7 @@
 		
 		//===========================수정/삭제 화면==========================
 		//수정 - 저장 버튼
-		$(".saveBtn").on("click", function(e){alert($("form textarea[name='content']").val());
+		$(".saveBtn").on("click", function(e){
 			$.ajax({
 				url:"/job/free/modify",
 				type:"post",
@@ -289,22 +315,14 @@
 						location.reload();
 					},1000);
 				}else if(result == "fail"){
-					showAlert('수정중 오류가 발생하였습니다', 'fail');
+					showAlert('수정중 오류가 발생하였습니다', 'error');
 				}						
 			});//done
 		});
 		
 		//수정 - 삭제 버튼
 		$(".delBtn").on("click", function(e){
-			swal({
-			  title: '해당 게시물을 삭제하시겠습니까?',
-			  type: 'warning',
-			  showCancelButton: true,
-			  confirmButtonColor: '#3085d6',
-			  cancelButtonColor: '#d33',
-			  confirmButtonText: '예',
-				cancelButtonText: '아니오'
-			}).then(function () {
+			showConfirm('해당 게시물을 삭제하시겠습니까?', function () {
 				$.ajax({
 					url:"/job/free/remove",
 					type:"post",
@@ -318,10 +336,10 @@
 							self.close();
 						},1000);
 					}else if(result == "fail"){
-						showAlert('삭제중 오류가 발생하였습니다', 'fail');
+						showAlert('삭제중 오류가 발생하였습니다', 'error');
 					}						
 				});//done
-			});//then
+			});	
 		});
 		
 		//수정 - 취소 버튼
